Complete CUDA implementation

+++ Core/Defines/Defines.cpp
@@ -59,7 +59,7 @@ namespace SingleNet
             float sum = 0.0f;
 
             for (size_t i = 0; i < y.size(); i++)
-                sum += std::pow(y[i] - y_[i], 2);
+                sum += std::powf(y[i] - y_[i], 2);
 
             return sum / y.size();
         };
